move classes in transactions.py to core.py, to align with parent class organization
