+++ README.md
@@ -18,7 +18,7 @@ pre-commit run --all-files`
 
 ```shell
 cd stac_fastapi/elasticsearch
-pip install .[dev]
+pip install -e .[dev]
 ```
 
 ## Building
