+++ docker-compose.yml
@@ -38,6 +38,6 @@ services:
       discovery.type: single-node
       network.host: 0.0.0.0
       http.port: 9200
-      ES_JAVA_OPTS: -Xms512m -Xmx512m
+      ES_JAVA_OPTS: -Xms512m -Xmx4g
     ports:
       - "9200:9200"
